# Add an experimental option to use fast backward in the hash encoder