--- a/scripts/train_360_v2_garden.sh
+++ b/scripts/train_360_v2_garden.sh
@@ -8,4 +8,4 @@ export DOWNSAMPLE=0.25 # to avoid OOM
 python3 train.py \
     --root_dir $ROOT_DIR/garden --dataset_name colmap \
     --exp_name garden --downsample $DOWNSAMPLE \
-    --scale 8.0
+    --scale 8.0 --batch_size 4096