This PR adds a new option `--half_opt` that enables the use of a faster backward in the hash encoder, which also allows for `float16` acceleration. 

### Speed up info (autodiff -> half_opt):
- lego: 164s -> 116.71s
- garden: 354s -> 223.23s